Move `ethlike/ethlikeutil` to `ethlike` package
Moved the tools from `ethlike/ethlikeutil` to
`ethlike` in order to simplify the package
structure and improve client code readability.
