+++ pkg/chain/ethereum/ethutil/rate_limiter_test.go
@@ -5,7 +5,7 @@ import (
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/keep-network/keep-common/pkg/chain/ethlike/ethlikeutil"
+	"github.com/keep-network/keep-common/pkg/chain/ethlike"
 	"math/big"
 	"strings"
 	"sync"
@@ -28,7 +28,7 @@ func TestRateLimiter(t *testing.T) {
 
 	rateLimitingClient := WrapRateLimiting(
 		client,
-		&ethlikeutil.RateLimiterConfig{
+		&ethlike.RateLimiterConfig{
 			RequestsPerSecondLimit: requestsPerSecondLimit,
 			ConcurrencyLimit:       concurrencyLimit,
 			AcquirePermitTimeout:   acquirePermitTimeout,
@@ -117,7 +117,7 @@ func TestRateLimiter_RequestsPerSecondLimitOnly(t *testing.T) {
 
 	rateLimitingClient := WrapRateLimiting(
 		client,
-		&ethlikeutil.RateLimiterConfig{
+		&ethlike.RateLimiterConfig{
 			RequestsPerSecondLimit: requestsPerSecondLimit,
 			ConcurrencyLimit:       concurrencyLimit,
 			AcquirePermitTimeout:   acquirePermitTimeout,
@@ -184,7 +184,7 @@ func TestRateLimiter_ConcurrencyLimitOnly(t *testing.T) {
 
 	rateLimitingClient := WrapRateLimiting(
 		client,
-		&ethlikeutil.RateLimiterConfig{
+		&ethlike.RateLimiterConfig{
 			RequestsPerSecondLimit: requestsPerSecondLimit,
 			ConcurrencyLimit:       concurrencyLimit,
 			AcquirePermitTimeout:   acquirePermitTimeout,
@@ -259,7 +259,7 @@ func TestRateLimiter_AcquirePermitTimout(t *testing.T) {
 
 	rateLimitingClient := WrapRateLimiting(
 		client,
-		&ethlikeutil.RateLimiterConfig{
+		&ethlike.RateLimiterConfig{
 			RequestsPerSecondLimit: requestsPerSecondLimit,
 			ConcurrencyLimit:       concurrencyLimit,
 			AcquirePermitTimeout:   acquirePermitTimeout,
