+++ pkg/chain/ethlike/block_counter.go
@@ -1,4 +1,4 @@
-package ethlikeutil
+package ethlike
 
 import (
 	"context"
