+++ pkg/chain/ethlike/logger.go
@@ -1,4 +1,4 @@
-package ethlikeutil
+package ethlike
 
 import "github.com/ipfs/go-log"
