+++ pkg/chain/ethlike/mining_waiter.go
@@ -1,4 +1,4 @@
-package ethlikeutil
+package ethlike
 
 import (
 	"context"
