+++ pkg/chain/ethlike/nonce_manager.go
@@ -1,4 +1,4 @@
-package ethlikeutil
+package ethlike
 
 import (
 	"context"
