+++ pkg/chain/ethlike/rate_limiter.go
@@ -1,4 +1,4 @@
-package ethlikeutil
+package ethlike
 
 import (
 	"context"
